better code layout for grpc implementations
