+++ src/core/controller/grpc-handler.ts
@@ -1,7 +1,5 @@
-import { BrowserConnectionInfo } from "../../shared/proto/browser";
-import { EmptyRequest } from "../../shared/proto/common";
 import { Controller } from "./index";
-import { getAllExtensionState } from "../storage/state";
+import { handleBrowserServiceRequest } from "./browser/index";
 import { ExtensionMessage } from "../../shared/ExtensionMessage";
 
 /**
@@ -27,7 +25,7 @@ export class GrpcHandler {
       // Handle BrowserService requests
       if (service === "cline.BrowserService") {
         return {
-          message: await this.handleBrowserServiceRequest(method, message),
+          message: await handleBrowserServiceRequest(this.controller, method, message),
           request_id: requestId
         };
       }
@@ -41,62 +39,6 @@ export class GrpcHandler {
     }
   }
 
-  /**
-   * Handle BrowserService requests
-   * @param method The method name
-   * @param message The request message
-   * @returns The response message
-   */
-  private async handleBrowserServiceRequest(method: string, message: any): Promise<any> {
-    switch (method) {
-      case "getBrowserConnectionInfo":
-        return this.getBrowserConnectionInfo(message);
-      default:
-        throw new Error(`Unknown method: ${method}`);
-    }
-  }
-
-  /**
-   * Get information about the current browser connection
-   * @param request The request message
-   * @returns The browser connection info
-   */
-  private async getBrowserConnectionInfo(request: EmptyRequest): Promise<BrowserConnectionInfo> {
-    try {
-      // Get browser settings from extension state
-      const { browserSettings } = await getAllExtensionState(this.controller.context);
-      
-      // Check if there's an active browser session by using the controller's handleWebviewMessage approach
-      // This is similar to what's done in controller/index.ts for the "getBrowserConnectionInfo" message
-      if (this.controller.task?.browserSession) {
-        // Access the browser session through the controller's task property
-        // Using indexer notation to access private property
-        const browserSession = this.controller.task.browserSession;
-        const connectionInfo = browserSession.getConnectionInfo();
-        
-        // Convert from BrowserSession.BrowserConnectionInfo to proto.BrowserConnectionInfo
-        return {
-          isConnected: connectionInfo.isConnected,
-          isRemote: connectionInfo.isRemote,
-          host: connectionInfo.host || "" // Ensure host is never undefined
-        };
-      }
-      
-      // Fallback to browser settings if no active browser session
-      return {
-        isConnected: false,
-        isRemote: !!browserSettings.remoteBrowserEnabled,
-        host: browserSettings.remoteBrowserHost || ""
-      };
-    } catch (error: unknown) {
-      console.error("Error getting browser connection info:", error);
-      return {
-        isConnected: false,
-        isRemote: false,
-        host: ""
-      };
-    }
-  }
 }
 
 /**
